[Arch] Extract posts feature into src/features/posts/
Move all post-related components (post-image, sandpack-editor,
transformer-copy-button, mdx-content) into src/features/posts/.
Extract PostsListPage and PostsDetailPage from route files,
making app/posts/ routes thin wrappers. Remove legacy mdx.tsx.
