+++ src/features/posts/index.ts
@@ -1,0 +1,2 @@
+export { PostsListPage } from "./pages/posts-list-page";
+export { PostsDetailPage } from "./pages/posts-detail-page";
